refactor: split seed client and loader
