+++ packages/cli/src/lib/services/seed/index.ts
@@ -1,3 +1,5 @@
 export * from './schema';
+export * from './interfaces';
 export * from './seed-client';
+export * from './seed-loader';
 export * from './id-mapper-client';
